requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
anthropic>=0.25.0  # or openai>=1.0.0
requests>=2.31.0
ckanapi>=4.7
openpyxl>=3.1.0
xlrd>=2.0.1
python-dotenv>=1.0.0
plotly>=5.17.0
diskcache>=5.6.3